requirements.txt
futures==3.1.1
grpcio==1.35.0
grpcio-tools==1.35.0
protobuf==3.14.0
click==7.1.2
pytest==6.2.2
pytest-benchmark==3.2.3
